Add git remote group sub-command
Remote groups are an existing feature of git presently managed
exclusively via the config. This patch gives them their own subcommand
for listing the existing groups with the intent of adding future
subcommands under it for adding/updating/removing groups.

As a side-effect this command can now be used in bash tab completion for
the fetch command which did not offer them as a completion option
previously. Note that the push command does not accept remote groups as
an argument.

Signed-off-by: Lazar Sumar <bugzilla@lazar.co.nz>